Apply bunch of updates including:
- Added a new extension method
There is an addition of new extention method `optional` which takes
two multiple parameters. It first checks whether or not the provided
input is null if its not then it applies the second validation rule
that is provided as the first parameter with rest of the params as
its arguments.

Pretty helpful when you want to apply validation only when the value
is not null, empty or undefined.

- Converted from object to array
Returning items is now an array rather than obnoxious index based
object.